Moved tests near implementation
